requirements.txt
numpy==1.22.3
tensorflow-cpu==2.9.1
opencv-python-headless
flask==2.1.2
flask-cors==3.0.10
gunicorn==20.1.0
